+++ openstates/importers/base.py
@@ -513,7 +513,7 @@ class BaseImporter:
 
         return data
 
-    def get_seen_sessions(self) -> typing.List[int]:
+    def get_seen_sessions(self) -> typing.List[str]:
         return [s.id for s in self.session_cache.values()]
 
     def resolve_person(